calibration ok

+++ include/EulerAngles.hpp
@@ -22,8 +22,8 @@ public:
     float pitch; // Inclinaison
     float roll;  // Roulis
 
-    uint32_t bno_timestamp; // Timestamp pour la validation des données
-    uint32_t esp_timestamp;     // Timestamp pour la validation des données
+    uint64_t bno_timestamp; // Timestamp pour la validation des données
+    uint64_t esp_timestamp;     // Timestamp pour la validation des données
     uint8_t sensorId;   // ID du capteur pour identifier la source des données
     float accuracy;     // Précision des données (si disponible)
     uint8_t precision;    // Précision des données (si disponible)
